internal: address app output variable as pointer
The output variable is transferred between the different functions of resources alignment checks and thus we need to have the option to update it anytime without the need to return it explicitly by the functions. Thus pass it's pointer to be able to update its value from anywhere.

Following this, update the unit test to use this variable instead of checking returned value, which is already contained in the "output" variable.

Signed-off-by: shajmakh <shajmakh@redhat.com>