internal: add IsAligned field to output struct
Numa field is not enough when alignment check is needed for multi resources. One simple example case is when the consumed resources are empty, Numa will be -1, which indicates the resources are not aligned to single numa. But that isn't true.
Add new boolean field, "IsAligned", to point out whether the resources are aligned or not.

Signed-off-by: shajmakh <shajmakh@redhat.com>